Moved test to termdbtest
