+++ server/test/tp/files/hg38/TermdbTest/tnse.txt
@@ -259,3 +259,7 @@ Ref	-101.8548145	72.46058642	HGNET_BCOR	HGNET_BCOR
 3178	121.1951911	73.75814818	PNET	ETMR
 3192	121.6732408	71.66798389	PNET	ETMR
 3206	-100.8863747	74.89473302	PNET	HGNET_BCOR
+3416	-101.8863747	75.89473302	PNET	HGNET_BCOR
+3444	-102.8863747	76.89473302	PNET	HGNET_BCOR
+
+
